+++ prism-core/src/main/java/network/darkhelmet/prism/core/storage/adapters/mariadb/MariaDbStorageAdapter.java
@@ -70,7 +70,7 @@ public class MariaDbStorageAdapter extends AbstractSqlStorageAdapter {
             SqlSchemaUpdater schemaUpdater,
             ISqlActivityQueryBuilderFactory queryBuilderFactory,
             CacheService cacheService,
-            @Named("serializerVersion") short serializerVersion,
+            @Named("serializerVersion") int serializerVersion,
             Path dataPath) {
         super(
             loggingService,
