+++ prism-core/src/main/java/network/darkhelmet/prism/core/storage/adapters/mysql/MysqlStorageAdapter.java
@@ -70,7 +70,7 @@ public class MysqlStorageAdapter extends AbstractSqlStorageAdapter {
             SqlSchemaUpdater schemaUpdater,
             ISqlActivityQueryBuilderFactory queryBuilderFactory,
             CacheService cacheService,
-            @Named("serializerVersion") short serializerVersion,
+            @Named("serializerVersion") int serializerVersion,
             Path dataPath) {
         super(
             loggingService,
