+++ prism-core/src/main/java/network/darkhelmet/prism/core/storage/adapters/postgres/PostgresStorageAdapter.java
@@ -61,7 +61,7 @@ public class PostgresStorageAdapter extends AbstractSqlStorageAdapter {
             SqlSchemaUpdater schemaUpdater,
             ISqlActivityQueryBuilderFactory queryBuilderFactory,
             CacheService cacheService,
-            @Named("serializerVersion") short serializerVersion,
+            @Named("serializerVersion") int serializerVersion,
             Path dataPath) {
         super(
             loggingService,
